Relocate experiment code from common-elements to give-before-pdf
