+++ src/app/pages/details/common/get-this-title-files/give-before-pdf/give-before-online.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import ThankYou, {useOnThankYouClick} from './thank-you-form';
+import useGiveLinks from './use-give-links';
 import CommonElements from './common-elements';
 import type {DonationPopupData} from './use-donation-popup-data';
 
@@ -13,6 +14,7 @@ export default function GiveBeforeOnline({
     data: DonationPopupData;
 }) {
     const {showThankYou, onThankYouClick} = useOnThankYouClick();
+    const [controlLink] = useGiveLinks();
 
     if (showThankYou) {
         return <ThankYou link={link} close={close} />;
@@ -24,7 +26,11 @@ export default function GiveBeforeOnline({
             data-analytics-view
             data-analytics-nudge='donate'
         >
-            <CommonElements data={data} onThankYouClick={onThankYouClick} />
+            <CommonElements
+                data={data}
+                onThankYouClick={onThankYouClick}
+                giveLink={controlLink}
+            />
             <a href={link} onClick={close} className='btn go-to'>
                 Go to your book
             </a>
